流式输出

+++ src/main/java/com/dyx/pojo/ChatForm.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class ChatForm {
-    private int memoryId;//对话id
+    private Long memoryId;//对话id
     private String message;//用户问题
 }
